#454: improve start* commands
