+++ packages/gluestick/src/renderer/index.js
@@ -19,12 +19,27 @@ require('babel-register')(Object.assign({
 }, babelAdditions));
 
 // Prepare logger.
+const loggerFactory = type => {
+  if (process.env.NODE_ENV === 'production') {
+    return value => {
+      const log = `${type.toUpperCase()}: ${value}`;
+      if (type === 'error') {
+        process.stderr.write(log);
+      } else {
+        process.stdout.write(log);
+      }
+    };
+  }
+  return value => {
+    process.send({ type, value });
+  };
+};
 const logger = {
-  info: value => { process.send({ type: 'info', value }); },
-  success: value => { process.send({ type: 'success', value }); },
-  error: value => { process.send({ type: 'error', value }); },
-  warn: value => { process.send({ type: 'warn', value }); },
-  debug: value => { process.send({ type: 'debug', value }); },
+  info: loggerFactory('info'),
+  success: loggerFactory('success'),
+  error: loggerFactory('error'),
+  warn: loggerFactory('warn'),
+  debug: loggerFactory('debug'),
 };
 
 // Exec renderer server.
